refactor: app 디렉토리 분리
