+++ src/app/app.tsx
@@ -1,0 +1,7 @@
+import AppRoutes from "./app-routes.tsx"
+
+const App = () => {
+  return <AppRoutes />
+}
+
+export default App
