+++ src/app/ui/globalLayout.tsx
@@ -1,0 +1,17 @@
+import { Outlet } from "react-router-dom"
+import Header from "../../components/Header"
+import Footer from "../../components/Footer"
+
+const GlobalLayout = () => {
+  return (
+    <div className="flex flex-col min-h-screen">
+      <Header />
+      <main className="flex-grow container mx-auto px-4 py-8">
+        <Outlet />
+      </main>
+      <Footer />
+    </div>
+  )
+}
+
+export default GlobalLayout
